requirements.txt
Django==3.2.7
Pillow==9.0.0
Djangorestframework==3.11.0